refactor(P2.9): split moderation.py handler into package
The handler was 443 LOC covering mute/unmute, ban/unban, blacklist (add via
!black, !spam, !blacklist list + pagination, confirm/cancel/unblock
callbacks), and welcome. Splitting along command boundaries keeps each file
under 250 LOC and agents stop having to scroll the whole file to find a
specific command.

Layout:
- moderation/__init__.py — composes moderation_router from four sub-routers;
  re-exports handler functions for test backward-compat
- moderation/_common.py — reply_required_error, is_user_check_error helpers
- moderation/mute.py — !mute, !unmute
- moderation/ban.py — !ban, !unban
- moderation/blacklist.py — !black, !spam, !blacklist + callbacks
- moderation/welcome.py — !welcome

Tests: patch paths updated for test_moderation_handlers.py (moderation.other
→ moderation.mute.other). 640 passed, 2 skipped.
