+++ app/presentation/telegram/handlers/moderation/blacklist.py
@@ -1,3 +1,5 @@
+"""Blacklist-related commands and callbacks: !black, !spam, !blacklist + pagination + unblock."""
+
 from aiogram import Bot, Router, types
 from aiogram.exceptions import TelegramBadRequest
 from aiogram.filters import Command
@@ -6,13 +8,14 @@ from sqlalchemy.ext.asyncio import AsyncSession
 
 from app.core.logging import get_logger
 from app.core.text import escape_html
-from app.db.repositories import (
-    ChatRepository,
-    MessageRepository,
-)
+from app.db.repositories import MessageRepository
 from app.moderation import blacklist as moderation_services
 from app.moderation import spam_service
 from app.moderation.user_service import UserService
+from app.presentation.telegram.handlers.moderation._common import (
+    is_user_check_error,
+    reply_required_error,
+)
 from app.presentation.telegram.utils import BlacklistConfirm, BlacklistPagination, UnblockUser, other
 from app.presentation.telegram.utils.blacklist import (
     build_blacklist_keyboard,
@@ -21,179 +24,13 @@ from app.presentation.telegram.utils.blacklist import (
     build_user_details_text,
 )
 
-logger = get_logger("handler.moderation")
-
-moderation_router = Router()
+logger = get_logger("handler.moderation.blacklist")
+router = Router()
 
 _BLACKLIST_PAGE_SIZE = 10
 
 
-def reply_required_error(action: str) -> str:
-    """Standard error when a command should be a reply."""
-    return f"Примените команду ответом на сообщение пользователя, которого нужно {action}. 🙏"
-
-
-def is_user_check_error() -> str:
-    """Standard error when target message does not contain a user."""
-    return "🚫 Это не пользователь или что-то пошло не так."
-
-
-@moderation_router.message(Command("mute", prefix="!/"))
-async def mute_user(message: types.Message, bot: Bot) -> None:
-    # Ensure command is used as a reply
-    if not message.reply_to_message:
-        await message.answer(reply_required_error("замутить"))
-        await message.delete()
-        return
-
-    if not message.reply_to_message.from_user:
-        await message.answer(is_user_check_error())
-        return
-
-    mute_guide = (
-        "Для мута пользователя используйте команду в формате:\n\n"
-        "<code>!mute [время] [единица времени]</code>\n\n"
-        "Примеры:\n<code>!mute 5m</code> - на 5 минут\n"
-        "<code>!mute 1h</code> - на 1 час\n"
-        "<code>!mute 1d</code> - на 1 день\n"
-        "<code>!mute 1w</code> - на 1 неделю"
-    )
-
-    # Parse and validate mute duration
-    try:
-        if not message.text:
-            await message.answer("Команда не может быть пустой.")
-            return
-        mute_duration = other.calculate_mute_duration(message.text)
-    except Exception:
-        answer = await message.answer(f"Мне не удалось распознать время мута!\n\n{mute_guide}")
-        await message.delete()
-        other.sleep_and_delete(answer, 10)
-        return
-
-    # Set permissions to mute the user
-    read_only_permissions = types.ChatPermissions(
-        can_send_messages=False,
-        can_send_media_messages=False,
-        can_send_polls=False,
-        can_send_other_messages=False,
-    )
-
-    try:
-        await bot.restrict_chat_member(
-            message.chat.id,
-            message.reply_to_message.from_user.id,
-            permissions=read_only_permissions,
-            until_date=mute_duration.until_date,
-        )
-        mention = other.get_user_mention(message.reply_to_message.from_user)
-        text_mute = (
-            f"{mention} в муте на {mute_duration.time} {mute_duration.unit}!\n\n"
-            f"Дата размута: {mute_duration.formatted_until_date()}"
-        )
-        await message.reply_to_message.reply(text_mute)
-        await message.delete()
-
-    except Exception as err:
-        await message.answer("Произошла ошибка. Попробуйте позже.")
-        logger.error(
-            "mute_failed",
-            error=str(err),
-            user_id=message.reply_to_message.from_user.id,
-            chat_id=message.chat.id,
-        )
-
-
-@moderation_router.message(Command("unmute", prefix="!/"))
-async def unmute_user(message: types.Message) -> None:
-    if not message.reply_to_message:
-        await message.answer(reply_required_error("размутить"))
-        await message.delete()
-        return
-
-    if not message.reply_to_message.from_user:
-        await message.answer(is_user_check_error())
-        return
-
-    # Reset permissions to default chat permissions
-    default_permissions = message.chat.permissions
-    if not default_permissions:
-        await message.answer("Похоже, что я нахожусь не в чате или у меня нет прав администратора.")
-        return
-
-    try:
-        await message.chat.restrict(
-            user_id=message.reply_to_message.from_user.id,
-            permissions=default_permissions,
-            until_date=0,
-        )
-        mention = other.get_user_mention(message.reply_to_message.from_user)
-        await message.answer(f"Пользователь {mention} размучен!")
-    except Exception as err:
-        await message.answer("Произошла ошибка. Попробуйте позже.")
-        logger.error(
-            "unmute_failed",
-            error=str(err),
-            user_id=message.reply_to_message.from_user.id,
-            chat_id=message.chat.id,
-        )
-
-
-@moderation_router.message(Command("ban", prefix="!/"))
-async def ban_user(message: types.Message, bot: Bot) -> None:
-    if not message.reply_to_message:
-        await message.answer(reply_required_error("забанить"))
-        return
-
-    if not message.reply_to_message.from_user:
-        await message.answer(is_user_check_error())
-        return
-
-    try:
-        await bot.ban_chat_member(message.chat.id, message.reply_to_message.from_user.id)
-        mention = other.get_user_mention(message.reply_to_message.from_user)
-        await message.answer(f"Пользователь {mention} забанен")
-    except Exception as err:
-        error_msg = await message.answer("Что-то пошло не так. Попробуйте позже.")
-        logger.error(
-            "ban_failed",
-            error=str(err),
-            user_id=message.reply_to_message.from_user.id,
-            chat_id=message.chat.id,
-        )
-        other.sleep_and_delete(error_msg, 10)
-
-    await message.delete()
-
-
-@moderation_router.message(Command("unban", prefix="!/"))
-async def unban_user(message: types.Message, bot: Bot) -> None:
-    if not message.reply_to_message:
-        await message.answer(reply_required_error("разбанить"))
-        return
-
-    if not message.reply_to_message.from_user:
-        await message.answer(is_user_check_error())
-        return
-
-    try:
-        await bot.unban_chat_member(message.chat.id, message.reply_to_message.from_user.id)
-        mention = other.get_user_mention(message.reply_to_message.from_user)
-        await message.answer(f"Пользователь {mention} разбанен")
-    except Exception as err:
-        error_msg = await message.answer("Что-то пошло не так. Попробуйте позже.")
-        logger.error(
-            "unban_failed",
-            error=str(err),
-            user_id=message.reply_to_message.from_user.id,
-            chat_id=message.chat.id,
-        )
-        other.sleep_and_delete(error_msg, 10)
-
-    await message.delete()
-
-
-@moderation_router.message(Command("black", prefix="!/"))
+@router.message(Command("black", prefix="!/"))
 async def full_ban(message: types.Message, message_repo: MessageRepository, db: AsyncSession) -> None:
     if not message.reply_to_message:
         await message.answer(reply_required_error("добавить в черный список"))
@@ -235,7 +72,7 @@ async def full_ban(message: types.Message, message_repo: MessageRepository, db: 
     await message.delete()
 
 
-@moderation_router.message(Command("spam", prefix="!"))
+@router.message(Command("spam", prefix="!"))
 async def label_spam(message: types.Message, message_repo: MessageRepository, db: AsyncSession) -> None:
     if not message.reply_to_message:
         answer = await message.answer(reply_required_error("пометить как спам"))
@@ -276,19 +113,7 @@ async def label_spam(message: types.Message, message_repo: MessageRepository, db
     await message.delete()
 
 
-@moderation_router.message(Command("welcome", prefix="!/"))
-async def welcome_change(message: types.Message, chat_repo: ChatRepository) -> None:
-    if not message.text:
-        await message.answer("Сообщение не может быть пустым.")
-        return
-    welcome_message = message.text.partition(" ")[2]
-    await chat_repo.update_welcome_message(message.chat.id, welcome_message)
-    await message.answer("<b>Приветственное сообщение изменено!</b>")
-    await message.answer(welcome_message)
-    await message.delete()
-
-
-@moderation_router.callback_query(BlacklistConfirm.filter())
+@router.callback_query(BlacklistConfirm.filter())
 async def process_blacklist_confirm(
     callback: types.CallbackQuery,
     callback_data: BlacklistConfirm,
@@ -330,19 +155,18 @@ async def process_blacklist_confirm(
     await callback.answer()
 
 
-@moderation_router.callback_query(lambda c: c.data == "cancel_blacklist")
+@router.callback_query(lambda c: c.data == "cancel_blacklist")
 async def process_blacklist_cancel(callback: types.CallbackQuery) -> None:
     if callback.message and isinstance(callback.message, types.Message):
         await callback.message.edit_text("Действие отменено")
     await callback.answer()
 
 
-@moderation_router.message(Command("blacklist", prefix="!/"))
+@router.message(Command("blacklist", prefix="!/"))
 async def show_blacklist(message: types.Message, user_service: UserService) -> None:
     """Show blacklist with pagination or search for specific user."""
     command_args = message.text.split()[1:] if message.text else []
 
-    # Search for specific user if argument provided
     if command_args:
         identifier = command_args[0]
         user = await user_service.find_blocked_user(identifier)
@@ -352,7 +176,6 @@ async def show_blacklist(message: types.Message, user_service: UserService) -> N
             await message.delete()
             return
 
-        # Show individual user details
         text = build_user_details_text(user)
         keyboard = build_user_details_keyboard(user)
 
@@ -360,7 +183,6 @@ async def show_blacklist(message: types.Message, user_service: UserService) -> N
         await message.delete()
         return
 
-    # Show paginated blacklist
     await _show_blacklist_page(message, user_service, page=0)
 
 
@@ -376,8 +198,6 @@ async def _show_blacklist_page(
         return
 
     total_pages = (len(blocked_users) + _BLACKLIST_PAGE_SIZE - 1) // _BLACKLIST_PAGE_SIZE
-
-    # Ensure page is within bounds
     page = max(0, min(page, total_pages - 1))
 
     text = build_blacklist_text(len(blocked_users), page, total_pages, _BLACKLIST_PAGE_SIZE, query)
@@ -387,7 +207,7 @@ async def _show_blacklist_page(
     await message.delete()
 
 
-@moderation_router.callback_query(BlacklistPagination.filter())
+@router.callback_query(BlacklistPagination.filter())
 async def handle_blacklist_pagination(
     callback: types.CallbackQuery,
     callback_data: BlacklistPagination,
@@ -401,8 +221,6 @@ async def handle_blacklist_pagination(
 
     blocked_users = await user_service.get_blocked_users()
     total_pages = (len(blocked_users) + _BLACKLIST_PAGE_SIZE - 1) // _BLACKLIST_PAGE_SIZE
-
-    # Ensure page is within bounds
     page = max(0, min(callback_data.page, total_pages - 1))
 
     text = build_blacklist_text(len(blocked_users), page, total_pages, _BLACKLIST_PAGE_SIZE, callback_data.query)
@@ -411,11 +229,10 @@ async def handle_blacklist_pagination(
     try:
         await callback.message.edit_text(text, reply_markup=keyboard.as_markup())
     except TelegramBadRequest:
-        # Fallback to sending new message if edit fails
         await callback.message.answer(text, reply_markup=keyboard.as_markup())
 
 
-@moderation_router.callback_query(UnblockUser.filter())
+@router.callback_query(UnblockUser.filter())
 async def unblock_user_callback(
     callback: types.CallbackQuery,
     callback_data: UnblockUser,
@@ -435,7 +252,6 @@ async def unblock_user_callback(
         user = member.user
         user_identifier = user.username or user.first_name or str(user.id)
     except Exception:
-        # Fallback: user not found, use ID as identifier
         user_identifier = str(user_id)
 
     if callback.message and isinstance(callback.message, types.Message):
